Made warning and error integration work. Converted camera and tile method to be simpler. Scaling works completely differently. Converted a lot of things to ints for speeds.

+++ include/tile.h
@@ -14,8 +14,8 @@ typedef struct Tile {
     SDL_Texture* tile_texture; /*!< a pointer to an SDL_Texture */
     SDL_Texture* sprite_texture; /*!< a pointer to am SDL_Texture */
     int tile_id;
-    float x; /*!< x position of tile. */
-    float y; /*!< y position of tile. */
+    int x; /*!< x position of tile. */
+    int y; /*!< y position of tile. */
 } Tile;
 
 #endif
